Implement basic functionality of mRPC (#92)
* save

* move transport implementations to a separate directory

* save

* add mrpc register (not finished)

* save, working on socket datapath

* save, creating service and customer

* save, pass compile

* minor fix, update KOALA_PATH

* save, graph abstraction

* save, add infrastructure of building engine graph

* save, update customer and server, a lot compile error to fix

* pass compile

* minor fix

* fix minor bug in print_lat

* save

* save

* save, implementing uverbs

* save, port ulib to koala

* save

* save, implement rpc_adapter

* save, fix several issues, pass compile

* finish coding, there are many places which do not make much sense

* fix some bugs

* minor update

* minor changes, also update a todo list for the initial RPC to run

* finish rpc client side async glue code and non-blocking get_request in rpc_adapter

* implemented switch_addr_space at the koala side

* fix many bugs

* fix more bugs, now can do rpc of a round-trip

* fix many bugs, now switch_addr_space works

* now can send multiple RPC requests

* save

* fix many bugs in memory mapping, now can stably run multiple messages

* minor change

* backport bugfix from the main branch

* minor change

+++ src/rdma/src/ibv.rs
@@ -447,7 +447,7 @@ impl<'ctx> CompletionQueue<'ctx> {
 
     /// Returns the capacity of the CQ. The CQ can hold have at least this number of elements.
     #[inline]
-    pub fn capacity(&self) -> usize {
+    pub fn capacity(&self) -> u32 {
         unsafe { &*self.cq }.cqe as _
     }
 }
